requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
more-itertools==4.3.0
numpy==1.15.2
pkg-resources==0.0.0
pluggy==0.7.1
py==1.6.0
pydash==4.7.3
pyenchant==2.0.0
pytest==3.8.1
six==1.11.0
